split lib.rs into separate files
