+++ src/connector.rs
@@ -1,0 +1,47 @@
+use std::fmt;
+use std::marker::Unpin;
+
+use futures_io::{AsyncRead, AsyncWrite};
+use native_tls::{Error};
+
+use crate::TlsStream;
+use crate::handshake::handshake;
+
+/// A wrapper around a `native_tls::TlsConnector`, providing an async `connect`
+/// method.
+#[derive(Clone)]
+pub struct TlsConnector(native_tls::TlsConnector);
+
+impl TlsConnector {
+    /// Connects the provided stream with this connector, assuming the provided
+    /// domain.
+    ///
+    /// This function will internally call `TlsConnector::connect` to connect
+    /// the stream and returns a future representing the resolution of the
+    /// connection operation. The returned future will resolve to either
+    /// `TlsStream<S>` or `Error` depending if it's successful or not.
+    ///
+    /// This is typically used for clients who have already established, for
+    /// example, a TCP connection to a remote server. That stream is then
+    /// provided here to perform the client half of a connection to a
+    /// TLS-powered server.
+    pub async fn connect<S>(&self, domain: &str, stream: S) -> Result<TlsStream<S>, Error>
+    where
+        S: AsyncRead + AsyncWrite + Unpin,
+    {
+        handshake(move |s| self.0.connect(domain, s), stream).await
+    }
+}
+
+impl fmt::Debug for TlsConnector {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("TlsConnector").finish()
+    }
+}
+
+impl From<native_tls::TlsConnector> for TlsConnector {
+    fn from(inner: native_tls::TlsConnector) -> TlsConnector {
+        TlsConnector(inner)
+    }
+}
+
